# agri-frontend/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "AgriSmart Benin API",
  "description": "Backend Monolithique Modulaire pour le Challenge Agriculture Intelligente",
  "version": "4.0.0"
 },
 "paths": {
...
  "/api/v1/market/offers": {
   "post": {
    "tags": [
     "Marché Agricole"
    ],
    "summary": "Publish Offer",
    "operationId": "publish_offer_api_v1_market_offers_post",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/HarvestOfferCreate"
       }
      }
     }
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HarvestOfferOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "get": {
    "tags": [
     "Marché Agricole"
    ],
    "summary": "Catalogue public des offres actives",
    "operationId": "list_market_offers_api_v1_market_offers_get",
    "parameters": [
     {
      "name": "commune",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "maxLength": 60
        },
        {
         "type": "null"
        }
       ],
       "title": "Commune"
      }
     },
     {
      "name": "department",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "maxLength": 60
        },
        {
         "type": "null"
        }
       ],
       "title": "Department"
      }
     },
     {
      "name": "product",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "maxLength": 100
        },
        {
         "type": "null"
        }
       ],
       "title": "Product"
      }
     },
     {
      "name": "max_price",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "number",
         "exclusiveMinimum": 0
        },
        {
         "type": "null"
        }
       ],
       "title": "Max Price"
      }
     },
     {
      "name": "skip",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "minimum": 0,
       "default": 0,
       "title": "Skip"
      }
     },
     {
      "name": "limit",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 100,
       "minimum": 1,
       "default": 20,
       "title": "Limit"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/HarvestOfferOut"
         },
         "title": "Response List Market Offers Api V1 Market Offers Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/market/offers/me": {
   "get": {
    "tags": [
     "Marché Agricole"
    ],
    "summary": "Mes offres (tous statuts)",
    "operationId": "my_offers_api_v1_market_offers_me_get",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "$ref": "#/components/schemas/HarvestOfferOut"
         },
         "type": "array",
         "title": "Response My Offers Api V1 Market Offers Me Get"
        }
       }
      }
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ]
   }
  },
  "/api/v1/market/offers/{offer_id}": {
   "get": {
    "tags": [
     "Marché Agricole"
    ],
    "summary": "Get Offer",
    "operationId": "get_offer_api_v1_market_offers__offer_id__get",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "offer_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Offer Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HarvestOfferOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   },
   "patch": {
    "tags": [
     "Marché Agricole"
    ],
    "summary": "Update Offer",
    "operationId": "update_offer_api_v1_market_offers__offer_id__patch",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "offer_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Offer Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/HarvestOfferUpdate"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HarvestOfferOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/market/offers/{offer_id}/status": {
   "patch": {
    "tags": [
     "Marché Agricole"
    ],
    "summary": "Marquer vendue, retirer ou réactiver",
    "operationId": "update_offer_status_api_v1_market_offers__offer_id__status_patch",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "offer_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Offer Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/OfferStatusUpdate"
       }
      }
     }
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HarvestOfferOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/market/offers/{offer_id}/interest": {
   "post": {
    "tags": [
     "Marché Agricole"
    ],
    "summary": "Signaler son intérêt au producteur (il est notifié)",
    "operationId": "express_interest_api_v1_market_offers__offer_id__interest_post",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "offer_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Offer Id"
      }
     }
    ],
    "requestBody": {
     "required": true,
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/InterestCreate"
       }
      }
     }
    },
    "responses": {
     "201": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/InterestOut"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/market/offers/{offer_id}/interests": {
   "get": {
    "tags": [
     "Marché Agricole"
    ],
    "summary": "Acheteurs intéressés par mon offre",
    "operationId": "list_interests_api_v1_market_offers__offer_id__interests_get",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "offer_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "string",
       "title": "Offer Id"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/InterestOut"
         },
         "title": "Response List Interests Api V1 Market Offers  Offer Id  Interests Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/market/prices": {
   "get": {
    "tags": [
     "Marché Agricole"
    ],
    "summary": "Prix de référence par produit et département (FCFA/kg)",
    "operationId": "reference_prices_api_v1_market_prices_get",
    "parameters": [
     {
      "name": "product",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "maxLength": 100
        },
        {
         "type": "null"
        }
       ],
       "title": "Product"
      }
     },
     {
      "name": "department",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "maxLength": 60
        },
        {
         "type": "null"
        }
       ],
       "title": "Department"
      }
     },
     {
      "name": "days",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 730,
       "minimum": 7,
       "default": 90,
       "title": "Days"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/market/price-suggestion": {
   "get": {
    "tags": [
     "Marché Agricole"
    ],
    "summary": "Fourchette de prix conseillée avant de publier (sans IA)",
    "operationId": "price_suggestion_api_v1_market_price_suggestion_get",
    "parameters": [
     {
      "name": "product",
      "in": "query",
      "required": true,
      "schema": {
       "type": "string",
       "minLength": 2,
       "maxLength": 100,
       "title": "Product"
      }
     },
     {
      "name": "department",
      "in": "query",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string",
         "maxLength": 60
        },
        {
         "type": "null"
        }
       ],
       "title": "Department"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/api/v1/market/offers/ai-draft": {
   "post": {
    "tags": [
     "Marché Agricole"
    ],
    "summary": "Rédiger une annonce à partir d'une photo et de quelques mots (IA)",
    "operationId": "offer_draft_api_v1_market_offers_ai_draft_post",
    "requestBody": {
     "content": {
      "multipart/form-data": {
       "schema": {
        "$ref": "#/components/schemas/Body_offer_draft_api_v1_market_offers_ai_draft_post"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ]
   }
  },
...
 },
 "components": {
  "schemas": {
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "HarvestOfferCreate": {
    "properties": {
     "product_name": {
      "type": "string",
      "maxLength": 100,
      "minLength": 2,
      "title": "Product Name",
      "examples": [
       "Ananas Pain de Sucre"
      ]
     },
     "quantity_kg": {
      "type": "number",
      "exclusiveMinimum": 0.0,
      "title": "Quantity Kg"
     },
     "unit_price_fcfa": {
      "type": "number",
      "exclusiveMinimum": 0.0,
      "title": "Unit Price Fcfa"
     },
     "contact_phone": {
      "type": "string",
      "title": "Contact Phone"
     },
     "land_id": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Land Id",
      "description": "Parcelle d'origine : renseigne la localisation et assure la traçabilité"
     },
     "location_commune": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Location Commune"
     },
     "department": {
      "anyOf": [
       {
        "$ref": "#/components/schemas/Department"
       },
       {
        "type": "null"
       }
      ]
     },
     "client_ref": {
      "anyOf": [
       {
        "type": "string",
        "pattern": "^[A-Za-z0-9_-]{8,64}$"
       },
       {
        "type": "null"
       }
      ],
      "title": "Client Ref",
      "description": "Identifiant unique généré par l'application (ex. UUID). Un envoi rejoué avec le même client_ref (file d'attente hors ligne) renvoie la ressource existante au lieu d'un doublon."
     }
    },
    "type": "object",
    "required": [
     "product_name",
     "quantity_kg",
     "unit_price_fcfa",
     "contact_phone"
    ],
    "title": "HarvestOfferCreate"
   },
   "HarvestOfferOut": {
    "properties": {
     "id": {
      "type": "string",
      "title": "Id"
     },
     "farmer_npi": {
      "type": "string",
      "title": "Farmer Npi"
     },
     "product_name": {
      "type": "string",
      "title": "Product Name"
     },
     "quantity_kg": {
      "type": "number",
      "title": "Quantity Kg"
     },
     "unit_price_fcfa": {
      "type": "number",
      "title": "Unit Price Fcfa"
     },
     "contact_phone": {
      "type": "string",
      "title": "Contact Phone"
     },
     "land_id": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Land Id"
     },
     "location_commune": {
      "type": "string",
      "title": "Location Commune"
     },
     "department": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Department"
     },
     "status": {
      "type": "string",
      "enum": [
       "active",
       "sold",
       "withdrawn"
      ],
      "title": "Status"
     },
     "sold_quantity_kg": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Sold Quantity Kg"
     },
     "sold_unit_price_fcfa": {
      "anyOf": [
       {
        "type": "number"
       },
       {
        "type": "null"
       }
      ],
      "title": "Sold Unit Price Fcfa"
     },
     "sold_at": {
      "anyOf": [
       {
        "type": "string",
        "format": "date-time"
       },
       {
        "type": "null"
       }
      ],
      "title": "Sold At"
     },
     "interest_count": {
      "type": "integer",
      "title": "Interest Count",
      "default": 0
     },
     "client_ref": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Client Ref"
     },
     "created_at": {
      "type": "string",
      "format": "date-time",
      "title": "Created At"
     },
     "updated_at": {
      "type": "string",
      "format": "date-time",
      "title": "Updated At"
     },
     "tel_url": {
      "type": "string",
      "title": "Tel Url",
      "description": "Lien d'appel direct (tel:)"
     },
     "whatsapp_url": {
      "type": "string",
      "title": "Whatsapp Url",
      "description": "Lien WhatsApp avec message pré-rempli"
     }
    },
    "type": "object",
    "required": [
     "id",
     "farmer_npi",
     "product_name",
     "quantity_kg",
     "unit_price_fcfa",
     "contact_phone",
     "location_commune",
     "status",
     "created_at",
     "updated_at",
     "tel_url",
     "whatsapp_url"
    ],
    "title": "HarvestOfferOut"
   },
   "HarvestOfferUpdate": {
    "properties": {
     "product_name": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 100,
        "minLength": 2
       },
       {
        "type": "null"
       }
      ],
      "title": "Product Name"
     },
     "quantity_kg": {
      "anyOf": [
       {
        "type": "number",
        "exclusiveMinimum": 0.0
       },
       {
        "type": "null"
       }
      ],
      "title": "Quantity Kg"
     },
     "unit_price_fcfa": {
      "anyOf": [
       {
        "type": "number",
        "exclusiveMinimum": 0.0
       },
       {
        "type": "null"
       }
      ],
      "title": "Unit Price Fcfa"
     },
     "contact_phone": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Contact Phone"
     }
    },
    "type": "object",
    "title": "HarvestOfferUpdate"
   },
...
   "OfferStatusUpdate": {
    "properties": {
     "status": {
      "type": "string",
      "enum": [
       "active",
       "sold",
       "withdrawn"
      ],
      "title": "Status"
     },
     "sold_quantity_kg": {
      "anyOf": [
       {
        "type": "number",
        "exclusiveMinimum": 0.0
       },
       {
        "type": "null"
       }
      ],
      "title": "Sold Quantity Kg",
      "description": "Par défaut : toute la quantité"
     },
     "sold_unit_price_fcfa": {
      "anyOf": [
       {
        "type": "number",
        "exclusiveMinimum": 0.0
       },
       {
        "type": "null"
       }
      ],
      "title": "Sold Unit Price Fcfa",
      "description": "Par défaut : le prix affiché"
     }
    },
    "type": "object",
    "required": [
     "status"
    ],
    "title": "OfferStatusUpdate"
   },
...
  },
  "securitySchemes": {
   "HTTPBearer": {
    "type": "http",
    "scheme": "bearer"
   }
  }
 }
}